review access security

+++ src/main/java/com/explosion204/wclookup/service/ReviewService.java
@@ -8,33 +8,53 @@ import com.explosion204.wclookup.model.entity.User;
 import com.explosion204.wclookup.model.repository.ReviewRepository;
 import com.explosion204.wclookup.model.repository.ToiletRepository;
 import com.explosion204.wclookup.model.repository.UserRepository;
+import com.explosion204.wclookup.security.util.AuthUtil;
 import com.explosion204.wclookup.service.dto.identifiable.ReviewDto;
 import com.explosion204.wclookup.service.pagination.PageContext;
 import com.explosion204.wclookup.service.pagination.PaginationModel;
 import com.explosion204.wclookup.service.validation.annotation.ValidateDto;
+import org.apache.commons.lang3.StringUtils;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
+import org.springframework.security.access.AccessDeniedException;
 import org.springframework.stereotype.Service;
+
+import static com.explosion204.wclookup.security.ApplicationAuthority.ADMIN;
 
 @Service
 public class ReviewService {
     private final ReviewRepository reviewRepository;
     private final ToiletRepository toiletRepository;
     private final UserRepository userRepository;
+    private final AuthUtil authUtil;
 
     public ReviewService(
             ReviewRepository reviewRepository,
             ToiletRepository toiletRepository,
-            UserRepository userRepository
+            UserRepository userRepository,
+            AuthUtil authUtil
     ) {
         this.reviewRepository = reviewRepository;
         this.toiletRepository = toiletRepository;
         this.userRepository = userRepository;
+        this.authUtil = authUtil;
     }
 
     public PaginationModel<ReviewDto> findAll(PageContext pageContext) {
         PageRequest pageRequest = pageContext.toPageRequest();
         Page<ReviewDto> page = reviewRepository.findAll(pageRequest)
+                .map(ReviewDto::fromReview);
+        return PaginationModel.fromPage(page);
+    }
+
+    public PaginationModel<ReviewDto> findByToiletId(long toiletId, PageContext pageContext) {
+        PageRequest pageRequest = pageContext.toPageRequest();
+
+        if (toiletRepository.findById(toiletId).isEmpty()) {
+            throw new EntityNotFoundException(Toilet.class);
+        }
+
+        Page<ReviewDto> page = reviewRepository.findByToiletId(toiletId, pageRequest)
                 .map(ReviewDto::fromReview);
         return PaginationModel.fromPage(page);
     }
@@ -69,6 +89,8 @@ public class ReviewService {
     public ReviewDto update(ReviewDto reviewDto) {
         Review review = reviewRepository.findById(reviewDto.getId())
                 .orElseThrow(() -> new EntityNotFoundException(Review.class));
+        User ownerUser = review.getUser();
+        checkAuthority(ownerUser.getId());
 
         if (reviewDto.getText() != null) {
             review.setText(reviewDto.getText());
@@ -85,6 +107,18 @@ public class ReviewService {
     public void delete(long id) {
         Review review = reviewRepository.findById(id)
                 .orElseThrow(() -> new EntityNotFoundException(Review.class));
+        User ownerUser = review.getUser();
+        checkAuthority(ownerUser.getId());
+
         reviewRepository.delete(review);
+    }
+
+    private void checkAuthority(long ownerUserId) {
+         long currentUserId = authUtil.getAuthenticatedUserId();
+
+         // only review owner or admin can update/delete
+         if (!authUtil.hasAuthority(ADMIN.name()) && currentUserId != ownerUserId) {
+             throw new AccessDeniedException(StringUtils.EMPTY);
+         }
     }
 }
